Move Whisper transcriber into package
