+++ scinoephile/audio/transcription/whisper/__init__.py
@@ -1,0 +1,13 @@
+#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
+#  and distributed under the terms of the BSD license. See the LICENSE file for details.
+"""Whisper transcription.
+
+Package hierarchy (modules may import from any above):
+* transcriber
+"""
+
+from __future__ import annotations
+
+from .transcriber import WhisperTranscriber
+
+__all__ = ["WhisperTranscriber"]
